new member field sample_rate added into WaveletTreeRRR

+++ src/libsds/include/BitSeqRRR.h
@@ -105,7 +105,7 @@ public:
 
 	/** construct a BitSeqRRR from a given BitStr with any type */
 	template<typename oIntType>
-	explicit BitSeqRRR(const BitStr<oIntType>& bstr, uint32_t sample_rate = DEFAULT_SAMPLE_RATE) : sample_rate(sample_rate) {
+	explicit BitSeqRRR(const BitStr<oIntType>& bstr, size_t sample_rate = DEFAULT_SAMPLE_RATE) : sample_rate(sample_rate) {
 		build_basic(bstr);
 		build_sampled();
 	}
